[cluster-test] Refactor execute_jobs function into a separate module
Summary
We want to use the execute_jobs function directly from various experiments and not just ClusterTestRunner

Test Plan
Compiles successfully
