+++ testsuite/cluster-test/src/lib.rs
@@ -9,6 +9,7 @@ pub mod log_prune;
 pub mod prometheus;
 pub mod slack;
 pub mod suite;
+pub mod thread_pool_executor;
 pub mod tx_emitter;
 
 pub mod util {
